Added decimal and id for categories

+++ dotnetLab1-main/Domain/Item.cs
@@ -4,7 +4,7 @@ public class Item
 {
     public int Id { get; set; }
     public string Name { get; set; } = String.Empty;
-    public double PricePerUnit { get; set; }
+    public decimal PricePerUnit { get; set; }
     public uint Amount { get; set; }
     public List<DateTime> SupplyDateTimes { get; set; } = new();
